Move row schema to own file
